use shared lib in basic example
